CP2000_Production/requirements.txt
# Core dependencies
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openpyxl==3.1.2
pandas==2.1.2
urllib3==1.26.15

# PDF and Image Processing
PyMuPDF==1.23.5
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# File monitoring
watchdog==3.0.0

# API and networking
requests==2.31.0
httplib2==0.22.0

# Excel handling
xlrd==2.0.1
xlwt==1.3.0

# Performance optimization
numpy==1.26.1
multiprocess==0.70.15